More fixes for predictable execution

+++ picoquic/picoquic.h
@@ -384,7 +384,7 @@ void picoquic_connection_id_callback_free_ctx(void * cnx_id_cb_data);
  * change the content or length of the packet.
  */
 typedef uint32_t(*picoquic_fuzz_fn)(void * fuzz_ctx, picoquic_cnx_t* cnx, uint8_t * bytes, 
-    size_t bytes_max, size_t length, uint32_t header_length);
+    size_t bytes_max, size_t length, size_t header_length);
 void picoquic_set_fuzz(picoquic_quic_t* quic, picoquic_fuzz_fn fuzz_fn, void * fuzz_ctx);
 
 /* Setting a cc log directory in order to create per connection packet traces.
